wxwidgets: fix cross arm64-windows*
Fix `wxwidgets` cross build on arm64-windows* by disabling the building
of gettext catalogs when cross-compiling. I attempted to set paths to
`msgfmt` and `msgmerge`, but invoking the host executables did not
create the output files.

Add host dependency on `gettext[tools]`, for when cross-compiling is
fixed correctly.

Set `wxrc` executable name to `wxrc-3.3`, for some reason this did not
fail on the other triplets. This may cause issues in the project, but I
think it's already handled.

Signed-off-by: Rafael Kitover <rkitover@gmail.com>